+++ backend/utils.py
@@ -11,7 +11,7 @@ def get_join_url(ultravox_payload):
     """
     # Get the logger that was set up in the main modules
     logger = logging.getLogger("plivo_server")
-
+    logger.info(f"System Prompt:\n {json.dumps(ultravox_payload)}")
     logger.info(f"Getting joinUrl with payload: {json.dumps(ultravox_payload)}")
 
     api_url = f"{ULTRAVOX_API_BASE_URL}/calls?enableGreetingPrompt=true"